Migrate InvoiceRequest amount to handle floats

+++ backend/src/main/java/com/ditod/acme/domain/invoice/dto/InvoiceRequest.java
@@ -4,5 +4,5 @@ import com.ditod.acme.domain.invoice.Status;
 
 import java.util.UUID;
 
-public record InvoiceRequest(UUID customerId, int amount, Status status) {
+public record InvoiceRequest(UUID customerId, double amount, Status status) {
 }
